refactor: fix common→domain layering with module moves
Import-path rewrites only, no code changes:
- landing-page/cta.ts → components/common/cta.ts (9 importers incl.
  common/BottomFAQ and domain/tickets)
- travel-guide/Reveal → components/common/reveal/ (6 of 7 importers
  were outside travel-guide)
- BloomingEthFlower, faq-bg.svg, AddToCalendarButton, VenueDetails
  → components/common/ (were imported by common/BottomFAQ, Footer,
  and the travel-guide page from domain/landing-page)
- LargeCallout → domain/about (about.tsx is its only consumer), taking
  its callout* SCSS along as large-callout.module.scss + bg-eth-repeat.png
- WritingText → components/common/ (used by about + supporters)
