+++ devcon/src/components/common/BottomFAQ.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import Image, { type StaticImageData } from 'next/image'
 import ReactMarkdown from 'react-markdown'
-import FaqBg from 'components/domain/landing-page/images/new/faq-bg.svg'
+import FaqBg from 'components/common/images/faq-bg.svg'
 import { Faq, FaqItem } from 'components/common/faq'
-import { BloomingEthFlower } from 'components/domain/landing-page/BloomingEthFlower'
+import { BloomingEthFlower } from 'components/common/BloomingEthFlower'
 import { Link } from 'components/common/link'
-import { ctaSecondary } from 'components/domain/landing-page/cta'
+import { ctaSecondary } from 'components/common/cta'
 import { ArrowRight } from 'lucide-react'
 import { useTranslations } from 'next-intl'
