+++ devcon/src/components/common/VenueDetails.tsx
