+++ devcon/src/components/common/WritingText.tsx
