+++ devcon/src/components/domain/about/large-callout.module.scss
@@ -1,0 +1,35 @@
+.callout {
+  padding: 120px 0;
+  text-align: center;
+  position: relative;
+  overflow: hidden;
+  background: transparent;
+}
+
+.callout-pattern {
+  position: absolute;
+  inset: 0;
+  opacity: 0.15;
+  pointer-events: none;
+  background-image: url('./images/bg-eth-repeat.png');
+  background-repeat: repeat-x;
+  background-position: center;
+  background-size: auto min(100%, 200px);
+}
+
+.callout-text {
+  font-family: var(--font-poppins), 'Poppins', sans-serif;
+  font-weight: 400;
+  font-size: clamp(20px, 2.2vw, 32px);
+  line-height: 1.5;
+  letter-spacing: -0.25px;
+  color: #7235ed;
+  text-align: center;
+  position: relative;
+  z-index: 1;
+  justify-content: center;
+}
+
+.callout-bold {
+  font-weight: 700;
+}
