+++ devcon/src/components/domain/landing-page/JoinTheEvent.tsx
@@ -4,8 +4,8 @@ import NextLink from 'next/link'
 import { ArrowRight } from 'lucide-react'
 import { useTranslations } from 'next-intl'
 import { useFeaturedWave, useTicketsCtaLabel } from 'hooks/useWaveStates'
-import { ctaPrimary } from './cta'
-import { Reveal, RevealGroup } from 'components/domain/travel-guide/Reveal'
+import { ctaPrimary } from 'components/common/cta'
+import { Reveal, RevealGroup } from 'components/common/reveal/Reveal'
 import CrescentMoons from './images/new/crescent-moons.svg'
 import JoinLearn from './images/new/join-learn.jpg'
 import JoinConnect from './images/new/join-connect.jpg'
