+++ devcon/src/pages/about.tsx
@@ -12,7 +12,7 @@ import About3 from 'assets/images/carousel/about/about-3.jpg'
 import About4 from 'assets/images/carousel/about/about-4.jpg'
 import HeroBackground from './past-events-hero.png'
 import { Link } from 'components/common/link'
-import { LargeCallout } from 'components/domain/landing-page/LargeCallout'
+import { LargeCallout } from 'components/domain/about/LargeCallout'
 import { VideoCard } from 'components/common/card/VideoCard'
 import SwipeToScroll from 'lib/components/event-schedule/swipe-to-scroll-native/SwipeToScroll'
 import { Clock4, History, Users, Flag, Globe, CirclePlay, ExternalLink, ArrowRight, ArrowUpRight } from 'lucide-react'
