+++ devcon/src/pages/supporters/index.tsx
@@ -9,7 +9,7 @@ import IconInstagram from 'assets/icons/instagram.svg'
 import IconTelegram from 'assets/icons/telegram.svg'
 import IconEmail from 'assets/icons/ui-email.svg'
 import InfiniteScroll from 'lib/components/infinite-scroll'
-import { WritingText } from 'components/domain/landing-page/WritingText'
+import { WritingText } from 'components/common/WritingText'
 import themes from '../themes.module.scss'
 import HeroBackground from './hero-bg.png'
 import JaaliPattern from 'assets/images/pages/ecosystem-jaali-left.svg'
