+++ devcon/src/pages/travel-guide/index.tsx
@@ -4,10 +4,10 @@ import Page from 'components/common/layouts/page'
 import { PageHero } from 'components/common/page-hero'
 import { SEO } from 'components/domain/seo'
 import { Snapshot } from 'components/common/snapshot'
-import { VenueDetails } from 'components/domain/landing-page/VenueDetails'
+import { VenueDetails } from 'components/common/VenueDetails'
 import { AccordionGroup } from 'components/domain/travel-guide/AccordionGroup'
 import { InfoTable } from 'components/domain/travel-guide/InfoTable'
-import { Reveal, RevealGroup } from 'components/domain/travel-guide/Reveal'
+import { Reveal, RevealGroup } from 'components/common/reveal/Reveal'
 import DC8Logo from 'assets/images/dc-8/dc8-logo.png'
 import { VideoPlayer } from 'components/domain/travel-guide/VideoPlayer'
 import JaaliBottom from './images/decor/jaali-bottom-border.svg'
